[atlas][TCE] Add the new fields to the evaluator
Summary:
Add the new fields to the evaluator.

The price on certain products is not an absolute price, there can be options.  One particular case is installment payments like here which is common on the purchase of phones.  To solve this problem we need to collect additional fields.  In the short term please do the following:
We should not use the existing field to provide incorrect information.  Hence, a product selling for $12 per month for 20 months should have the following fields:
price (this applies to both sale and list price) = null
installment_sale_price = $12 (New field)
number_of_installments = 20
installment_frequency = monthly

Test Plan:
```
🌊pzhang@~/code/evaluator$ ./gradlew test

BUILD SUCCESSFUL in 0s
4 actionable tasks: 4 up-to-date
```

Reviewers: heathvinicombe, jzhao, pauldb, hsung

Reviewed By: heathvinicombe, pauldb

Differential Revision: https://phabricator.pinadmin.com/D603155